git-svn-id: http://code.imagej.net/svn/imagej/trunk@2727 be23795f-a60e-40d4-ba19-055798872e57

+++ core/core-plugins/src/main/java/imagej/core/plugins/convolve/Convolve3x3Operation.java
@@ -109,7 +109,7 @@ public class Convolve3x3Operation {
 
 		/** at each new neighborhood reset it's value sum to 0 */
 		@Override
-		public void initializeNeighborhood(int[] position) {
+		public void initializeNeighborhood(long[] position) {
 			sum = 0;
 		}
